src/main/java/de/innovationhub/prox/modules/tag/application/tag/usecase/queries/FindPopularTagsHandler.java
package de.innovationhub.prox.modules.tag.application.tag.usecase.queries;

import de.innovationhub.prox.commons.stereotypes.ApplicationComponent;
import de.innovationhub.prox.modules.tag.domain.tag.Tag;
import de.innovationhub.prox.modules.tag.domain.tagcollection.TagCollectionRepository;
import java.util.List;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;

@ApplicationComponent
@RequiredArgsConstructor
public class FindPopularTagsHandler {

  private final TagCollectionRepository tagCollectionRepository;

  public Page<Tag> handle(Pageable pageable) {
    return tagCollectionRepository.findPopularTags(pageable);
  }
}
